move reflection tree build functions to builder module
